refactor: 코드 정리 및 TabItem 파일로 빼기
